edssharp: add another extension for the EDS file for CountLabel
The count label is required to compile against the CANopenNode v4
library since the definitions are checked by code like the following
in CANopen.c

    #if OD_CNT_NMT != 1
     #error OD_CNT_NMT from OD.h not correct!
    #endif

This problem is exhibited when there is no xpd file that has the
custom properties and there is only an EDS file available. The
StorageLocation is an example of setting the custom property for
the CO_storageGroup and this is adding a similar extension for
the CO_countLabel.

An example eds for NMT 0x1000 is shown below
    [1000]
    ParameterName=Device type
    ObjectType=0x7
    ;CountLabel=NMT
    ;StorageLocation=ROM
    DataType=0x0007
    AccessType=ro
    DefaultValue=0x00000000
    PDOMapping=0

Signed-off-by: Charles Hardin <charles.hardin@redoxblox.com>